Fixed tests for folder by tags endpoint.
